fix EnvUtil import & minor refactor
